move the Autocomplete, Dropdown, and SearchBox components into src/app-common/inputs to organize all the input components together
