+++ src/app-common/inputs/tests/SearchBox.test.js
